Test: CTS: ignore monitor failures when testing remote node fencing
When the RemoteStonithd test killed the remote daemon, the loss was detected
via lrmd_tls_dispatch(), so nothing was done with the recurring monitors behind
the connection (which is a questionable situation).

Now that it freezes the daemon, the connection timeout is detected via
lrmd_event_disconnect(), and the monitors behind the connection are immediately
failed. Ignore those failures.