Build: spec: Provides: not dissected for extra arch qualification
That means that with arch-qualified-only virtual Provides: (introduced
as of 73e2c94a3 for which this is a follow-up), the dependants would
need to track Requires:/Recommends: etc. in the same qualified form
unconditionally (which was not previously true in pacemaker packaging
itself as arranged with the said commit, which led to this discovery),
since only that would match string-wise.  Solution is to declare
Provides: for both forms in parallel, providing the depending packages
with the benefit of choice (or conversely, to prevent hard to debug
install transaction unresolvability issues; Ken did a great work
on this front).

See also:
https://bugzilla.redhat.com/show_bug.cgi?id=1659259