Log: scheduler: downgrade clone pre-allocation message
When allocating clone instances, we pre-allocate instances to their existing
location if possible. If not, we would previously log a notice about
"Pre-allocation failed", which might make this seem more significant than it
is (it's normal when the instance is moving). So, downgrade it to info,
and don't say it "failed".

This also includes some trivial refactoring for efficiency and clarity.