Fix: attrd: start new election if writer is lost
If a writer is shutting down when it receives an attribute update, it will not
write it to the CIB. Previously, a new election wouldn't be held until another
attribute needed to be written, which may not happen in a reasonable time (or
at all). Now, we trigger a new election when the writer leaves the cluster.

rhbz#1535221