Feature: tools: Rearrange cli_resource_execute.
The guts of this function are now part of a new public function,
cli_resource_execute_from_params.  This new function takes a hash table
of parameters and runs the validation.  It does not require a
pre-existing resource, which means it can be used to validate new
resources from the command line.

cli_resource_execute now operates in terms of the new function.  It
extracts all the parameters from an existing resource, puts them into a
hash table, and calls cli_resource_execute_from_params.

See rhbz#1637020