Build: spec: add hint about sbd compatibility
The libpe_status API for the pe_working_set_t data type changed incompatibly
in 2.0.1. sbd 1.4.0 supports the new API, so add a Conflicts line for versions
older than that. Of course, the library soname version change is the hard
barrier, but this provides another hint to someone preparing for an upgrade.