Test: cts: temporarily disable any enabled cluster serivces when running remote tests
Cluster nodes are reused as remote nodes in remote tests. If cluster
services were enabled at boot, in case the remote node got fenced, the
cluster node would join instead of the expected remote one. Meanwhile
pacemaker_remote would not be able to start. Depending on the chances,
the situations might not be able to be orchestrated gracefully any more.