Test: scheduler: massively speed-up (+ get progress) w/ GNU parallel
In practice, with Intel i7 quadcore (8 threads), the speed-up is by the
factor of 4+ (when not setting number of jobs explicitly, i.e., letting
GNU parallel figure this on its own).  This mitigates a very distracting
slowdown in the linear development workflow, since ~50 seconds compared
to more than 3.5 minutes previously makes up for quite some subjective
difference.  Apparently, some other test suites are eligible to receive
this extension as well (since a similar style is followed, would be
mostly a matter of copy-paste); the other development workflow
de-alienating further on the map is configure/build related
(especially having meson + ninja combo in mind).

Note that this speed-up is only applied when "parallel" program is
available (considering $PATH, no need to special-case when configuring
the build, since the full backward compatibility, strictly linear
execution, can be achieved with "-j 1" passed as [separate] arguments).