Feature: pengine: Add pe_eval_date_expression function.
This is similar to pe_test_date_expression, but with a more detailed
return type.  Instead of just a regular "did this time succeed or not",
it returns what side of a range the given time falls on, or if it
satisfies the eq/neq op, and so forth.  pe_test_date_expression is now
implemented in terms of this new function but interprets the result and
returns a boolean.