Refactor: get rid of redundancy/wasted effort incurred with crm_perror
Problem with crm_perror macro is that it bypasses the a priori made
decisions about how/where the log messages should go since it logs
also to stderr, which makes for redundant journal entries at best
(see also 32dc10b0c).

But the convention makes it clear that logging shall be configured
early on, and from that point, it's well enough to only use that
as the only sink.

It would be unfortunate to change the semantics of crm_perror, though,
so we roll out the new crm_log_perror one, which hopefully makes it
apparent that the intended scope of its use is the same as with
crm_log (i.e. once the logging is configured).  This macro is applied
everywhere but in logging.c, since those functions may be those that
actually contribute to setting the logging up.

Also added is a new semantic patch meant to guard correct use of
crm_log_perror vs. crm_perror in the future.