Build: auto-parallelize build by default when using checkout + Linux
In a sense, this is a finalization of 41e2bf203 establishing the grounds
years ago :-p ... and only small improvements are to be expected in the
build process alone (outside of the tests, which a subsequent commit is
going to give a crack to, partially) -- some completely opaque to the
build system (permanently caching the build intermediate outcomes, e.g.
ccache) or for systems with predominance of notoriously slow bash
as a system shell, with push for a more efficient alternative like dash
(subject of a parallel proposal).

This feature won't be enabled by default with a tree in a distributed
tarball, since -- intentionally -- the respective helper is under
/maint hierarchy.  Users and downstreams are welcome to pass "-j"
to "make" explictly, as usual.  Similarly, even if it could work,
it won't when not under Linux, for /proc/cpuinfo not being portable
(patches welcome if someone tries to build from checkouts on other
systems).