Low: executor: reschedule recurring stonith monitor if not connected
If a recurring stonith monitor fails in its first occurrence due to the stonith
connection being lost, reschedule the monitor.

This is accomplished by having lrmd_rsc_execute_stonith() call
stonith_action_complete() as normal in this case. Previously, it would set the
rc and status and call cmd_finalize(), which was equivalent except for not
rescheduling the monitor. However, with an error status, the client would not
know the monitor was effectively cancelled, so the state information could get
out of sync. Also, stonith_connection_failed() continues to reschedule monitors
in a similar situation, so it makes sense to be consistent.