High: libcrmcommon: complete interrupted live migrations correctly
8435966 (2.0.0) got the sense of compare_version() wrong when attempting to
remove the old behavior for recording partial live migrations, instead
re-introducing it.

This causes a regression in behavior when a transition containing a live
migration is aborted after the migrate_to has been initiated but before the
migrate_from has been initiated. In that case, it would schedule a full start
on the destination node rather than a migrate_from.