Refactor: various: use Pacemaker Remote terminology consistently
... except in the public API, which is unchanged for backward compatibility.

Guest nodes were originally referred to as "container remote nodes", which is
easily confused with the sense of "container" as in isolation. Remote nodes
were originally referred to as "baremetal remote nodes". Update internal naming
as much as possible.

"Remote node" originally was a generic term for "remote node or guest node".
Outside of the public API, effort has been taken to avoid this usage.

"Pacemaker Remote" is still the term referring to the general technology
(server- and remote-side component code) that creates remote nodes and guest
nodes. So, "Pacemaker Remote node" is sometimes still used in this generic
"remote node or guest node" sense. That is admittedly confusing when compared
to "remote node", but no better alternative is obvious.