Refactor: scheduler: simplify handling of restart ordering
Don't condition pe_action_implies() on the desired state not being already
present, because pe_action_implies() handles that.

Don't condition clearing first's pe_action_migrate_runnable on first being
required, because if it's optional it doesn't matter, and if (now or in the
future) optional can be changed to required later, it will actually be
important.