Fix: scheduler: one group stop shouldn't make another required
1.1.7's 8d2f237d reused pe_order_restart ("stop resource before stopping it")
for "stop later group member before stopping earlier group member".

pe_order_restart includes a check for an unrunnable 'then', because in a
restart, even if the start is unrunnable, we still want to perform the stop.
However this check does not make sense for group stop ordering, and as of
1.1.10, this caused a regression where a group member could be unnecessarily
stopped.

Example scenario: if a resource is ordered after a group member, and the
resource failed with on-fail=block, that would make the group member's
(optional) stop blocked as well, and that blocked stop would unnecessarily make
stops of later group members required.

This commit fixes the issue by only applying the check when the 'then' action
is a start. (RHBZ#1609453)