Feature: tools: Use formatted output in stonith_admin.
There's a lot going on in this commit, but it's kind of hard to break
down.  Here's the important stuff:

* Added --output-as= and --output-to= command line options via the
PCMK__OUTPUT_OPTIONS macro.

* fence_action_str is now provided by a custom message in libcrmcommon.

* Registration of the custom formatting stuff is a somewhat annoyingly
complicated block.  There's registration of supported formats, creating
the type, and then registration of extra messages.  Perhaps this can be
reduced somewhat.

* All text that doesn't get output in some formatted way is now written
to stderr so it can be split apart.  This includes an error message in
command line processing and some extra output in handle_history.

* -s output is now more structured.