Tests: cts-cli: simplify+fix regexp to catch crm_time_as_string's output
Previously, the cts-cli would only work with a zero-offset time zone
(suggesting that it was only run in the CI that may be expected to
have it like that).  Also drop the atypical enumeration of all digits
vs. range capture.

Backport-from-master-of: bfa3ddc527b8bfafe60b44185c1aa44abbdb5b7c