Build: xml: Strip subdirs in best-match.sh output.
The API result RNG files live in an api/ subdirectory.  In the future,
we may wish to move other RNG files into their own subdirectories.
However, the auto-generated RNG file that includes all the API sub-files
will also be written into the api/ subdirectory.  Thus, we should strip
the subdirectory out of the output so the reference is correct.