Med: libcrmcommon/iso8601: fix train of thinkos regarding time offsets
It's not very clear why the whole parsing is so diverted from the actual
ISO 8601 standard prescribed syntax, but multiple problems were
discovered when testing something as basic as 2019-04-25T21:42:54+0200:

1/ the proclaimed ISO 8601 date/time format parser is schizophrenic
   in a sense it both (a) supports undocumented extensions above what
   the standard defines (like a tolerance to white space at various
   places within the format string, like mentioned in 2/), i.e.,
   implementing a support for a superset of ISO 8601, and at the same
   time, (b) doesn't implement support for the mandated syntactical
   constructions (said optional support for white spaces is turned into
   requiring them where are undue per the standard, meaning that the
   above would be treated as "2019-04-25T21:42:54 of whatever currently
   configured timezone"), i.e., effectively implementing a subset of
   the standard

2/ with 1/ fixed, that is, at least at the location affecting parsing
   of the above example input, '+' sign (e.g. +0200) would consume
   artificially a position where digit was expected (the above would
   be garbled and parsed as if it effectively was +0020)

These may, apparently, derail the carefully tweaked execution plans
of the users.

* * *

The identified problems are fixed while retaining the original
tolerance (against the standard), and a respective tests to capture
any regressions relevant to these fixes is added (two variants of
the same test, actually, so as to limit the possiblity of the problem
marked 1/ being masked -- simply because the runner's timezone matches
that marked with an offset, as happened in the first iteration of this
fix, thanks to CI/Ken for noticing it :-)

Sooner rather than later, we should swap this highly problematic
chunk of code for some offloaded solution, Glib's GDateTime:
https://developer.gnome.org/glib/stable/glib-GDateTime.html#g-date-time-new-from-iso8601
being a prime target.

Such a change will likely need to be accompanied some schema
upgrade procedures as well, so as not keep users on their own
with configurations that once worked well (only because of an
unexpectedly botched and untested implementation).