Fix: libcrmcommon/iso8601: output actually valid ISO8601 date/time specs
Sadly, despite the standard references throughout, the outputs were
ostentatiously ignoring that and were producing non-conforming outputs
instead.  This is now tackled, and we also care to provide well-formed
inputs to cts-cli tests that are very related (note that -E switch to
iso8601 are meant for testing only, hence shall be rather un-publicized
at some point).

There was some fallout in terms of changed expected outputs for
cts-scheduler as well.