Build: 16% boost in configure time with dash in place of bash
Measurements taken with i7-6820HQ and updated Fedora Rawhide; the boost
concerns overall autoconf+configure time (ccache was present, if it had
any influence here, it would be uniform regardless of shell).

Bash is notoriously slow, but Fedora/EL distros use it as a default
/bin/sh, so apply a convenience trick directly in ./autogen.sh so that,
when invoked with "roll [CONFIGURE-ARGS]" as argument(s), it will
attempt to stick with dash (or "[busybox] ash", which is on par with
the speef of the former) if available.  Merely for completeness, it will
also enable a workaround for the unexpected situation there's no /bin/sh
on the system (or /usr/local/bin/sh or so is preferred for casual uses
per the PATH priorities?), as well as explicit restoring of bash
preference (using BUILD_SH=/bin/bash ./autogen roll ...).

In a sense, it restores some prior convenience arrangements present
till 19a1794d9, but the difference is that the new addition is terse
and very easy to grasp.

For the time being, .gitlab-ci.yml doesn't use it, as it it's
expected that the overhead of installing dash/busybox in addition
to the inherently present bash on Fedora will not gain anything,
quite the contrary.  It may be reconsidered in the context of
actual test executions, though a separate measurement is needed.

Note that also drop of "local" non-POSIX extension from configure
script was in order (dash warns about that).  Ironically enough,
doing so exposed pre-existing thinko (misused cc_{temp,restore}_flags}
since they are not re-entrant-within-scope, apparently) from 03331744d.
Finally, there's a little drawback with local-free shell when going
back to a historical commit, since any subsequent "make" will
trigger "configure", but using already defined CONFIG_SHELL, hence
falling on the nose (recovery is as simple as running the script
manually, only continuing with "make").