Maint: gnulib: breakdown fetch recipe, libtool switch & 00gnulib.m4 drop
This is a forerunner of a rather fat annex of new modules to borrow from
gnulib, so we'd rather start with cutting of what we expressly don't
require.  Note that md5.c code used to be consumed through a custom
recipes in lib/common/Makefile.am, rather than deferring to Gnulib's
own recipes, which is now changed as well.