Feature: tools: crm_resource: new option for "free form" date/time spec
In addition to pre-existing interface to specify the lifespan of the
newly created constraints as "period from now" (using -u), we proceed
to add a similar interface we've just added to crm_{rule,simulate}, that
is, ability to specify such end point in absolute terms (until date/time
X, using "--lifetime-until X"), and likewise, this specification can
be of arbitrary free form parseable with pcmk__time_new_versatile()
on systems where supported.

Also add a few tests to exercise this new input possibilities.