Maint: tools: unify help/man page for iso8601 tool wrt. references
In particular, unified with crm_{resource,rule,simulate} that
underwent such changes along teaching them to understand date/time free
forms (which is not appropriate in the context of this tool, it seems
-- that can be reconsidered later, though).

Also fix a handful of typos and cosmetic discrepancies, incl.
suggestions of actually ISO 8601 non-conforming date/time formats.