Maint: gnulib: prerequisites for parse-datetime & indirectly this module
As a consequence of:
https://lists.gnu.org/archive/html/bug-gnulib/2016-10/msg00045.html
we cannot consume parse-datetime module right away since we are strict
on LGPLv2+ license boundaries.  The solution for us is hence to obtain
at least the the so-licenced prerequisite modules by the standard
means of gnulib-tool as already established, and to grab its actual
constituting source (parse-datetime.[hy]) directly from checkout.
For the missing (non-LGPLv2+'d) prerequisite modules, we establish
reasonable surrogates (that's why those files pop up in te stats):

* strftime (strftime.[ch])
  - strftime.h surrogate header that defers directly to libc
    for strftime()

* timespec (timespec.[ch])
  - transitive dependency via gettime module, for which we inject
    gettime() wrapper over clock_gettime directly in parse-datetime.h
  - just an empty timespec.h is tracked then instead

* xalloc (xalloc.[ch])
  - xalloc.h surrogate header that defers to g_malloc/g_memdup

Finally, it's to be noted that the situation appeared to be untractable
like this after following monumental (in terms of seemingly untwistable
dependency on non-LGPLv2+'d module time_r) update to parse-datime:
https://lists.gnu.org/archive/html/bug-gnulib/2017-01/msg00133.html
so consuming the state of the source right prior to that change seemed
a pragmatic and good enough solution, modulo:

* a trivial quick fix we replay artifically with a sed one-linet:
  https://lists.gnu.org/archive/html/bug-gnulib/2017-04/msg00028.html

* similarly applied removal of a form-feed character from
  parse-datetime.y that would raise an alarm in Travis CI (MAINT=1)
  character sanity checking arrangement otherwise